二人目のモンスターの追加 モンスター専用の攻撃アニメーションを探せた お互いの攻撃当たり判定をほしい物同士でとることができる用になった

+++ src/System/UIBar.h
@@ -69,10 +69,10 @@ private:
 	int m_value_max;
 
 	//! valueの値が１減るたびにバーの減る量
-	 int value_decrease;
+	 float value_decrease;
 
 	// 今どれだけ増減してどれほぼ残っているか
-	int new_value;
+	float new_value;
 
 	//! 名前を保存するための変数
 	 char m_name[256] ;
